Rename Proofs/Subagent -> Proofs/Background (pure no-op rename)
