+++ crates/defra-agent/proofs/Proofs/Background.lean
@@ -1,0 +1,12 @@
+import Proofs.Background.State
+import Proofs.Background.Bridge
+import Proofs.Background.Transition
+import Proofs.Background.Properties
+import Proofs.Background.Executable
+
+/-!
+# Subagent
+
+Barrel import for subagent lifecycle (mode/policy state, BridgedState
+paired-context, bridge transitions, properties B1–B6).
+-/
