+++ crates/defra-agent/proofs/Proofs/Background/Executable.lean
@@ -1,4 +1,4 @@
-import Proofs.Subagent.Transition
+import Proofs.Background.Transition
 
 /-!
 # Subagent Executable Semantics
