--- crates/defra-agent/proofs/Proofs/Subagent.lean
@@ -1,12 +1,0 @@
-import Proofs.Subagent.State
-import Proofs.Subagent.Bridge
-import Proofs.Subagent.Transition
-import Proofs.Subagent.Properties
-import Proofs.Subagent.Executable
-
-/-!
-# Subagent
-
-Barrel import for subagent lifecycle (mode/policy state, BridgedState
-paired-context, bridge transitions, properties B1–B6).
--/
