+++ crates/defra-agent/src/background_tools.rs
