+++ crates/defra-agent/src/tool_call_lifecycle/recovery.rs
@@ -5,13 +5,13 @@ use chrono::{DateTime, Utc};
 use defra_node::EmbeddedNode;
 use serde::Deserialize;
 
-use crate::graphql::escape_graphql_string;
-use crate::interrupt::interrupt_request;
-use crate::session::execute_mutation_with_retry;
-use crate::subagent_tools::{
+use crate::background_tools::{
     fail_running_subagent_tool_call, load_parent_subagent_authorization, subagent_spawn_denial,
     subagent_tool_not_allowed_payload,
 };
+use crate::graphql::escape_graphql_string;
+use crate::interrupt::interrupt_request;
+use crate::session::execute_mutation_with_retry;
 
 use super::{
     subagent_request::create_subagent_request_with_request_id, AwaitMode, CancelPolicy,
